requirements.txt
httpx>=0.23,<1.0
nltk>=3.6
python-dotenv>=0.19
